Fixes for re-enabling more MSVC level 4 warnings: courgette/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/371153002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281696 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0ef486bd105ad367b055a3e9cc25cc67cf8d2a5d

+++ courgette/disassembler_elf_32.h
@@ -30,7 +30,7 @@ class DisassemblerElf32 : public Disassembler {
   // class encapsulates this behavior.  public for use in unit tests.
   class TypedRVA {
    public:
-    explicit TypedRVA(RVA rva) : rva_(rva), offset_(-1) {
+    explicit TypedRVA(RVA rva) : rva_(rva), offset_(static_cast<size_t>(-1)) {
     }
 
     virtual ~TypedRVA() { };
